Moved serialization file out of core directory
