+++ noether_gui/src/widgets/tpp_widget.cpp
@@ -3,7 +3,7 @@
 #include <noether_gui/widgets/configurable_tpp_pipeline_widget.h>
 #include <noether_gui/widgets/tpp_pipeline_widget.h>
 #include <noether_gui/utils.h>
-#include <noether_tpp/core/serialization.h>
+#include <noether_tpp/serialization.h>
 
 #include <pcl/io/vtk_lib_io.h>
 #include <QColorDialog>
